+++ CHANGES.md
@@ -4,6 +4,8 @@
  with a `org.apache.kafka.connect.data.Struct` value schema instead of as a `String`.
 - [NEW] Configuration option `cloudant.value.schema.struct.flatten` to flatten nested objects and
  arrays when using `cloudant.value.schema.struct=true`.
+ - [NEW] Configuration option `cloudant.omit.design.docs` to suppress message production for design
+ documents.
 - [BREAKING CHANGE] Changed build tooling to Gradle.
 - [NOTE] Version number format. The kafka-connect-cloudant version is no longer identical
  to an Apache Kafka version. The Apache Kafka version is appended to the kafka-cloudant-connect